Version 1.2 (IDE Optimization)
The version optimizes the API calls made to the LLM by using the concept of ITERATIVE DETECTION AND ESTIMATION. It implements Random Forest Regressor model that detects and estimates the best suitable prompting technique to generate the PR review with best possible accuracy score. With each new PR reviewed, the model gets better resulting in an increasing average accuracy score.